Solving old, not loaded at startup, transactions notification issue.
Problem: we are still signaling updates from really old transactions from the backend and, at the same time, the GUI is not loading more than 20k transactions into the transactionTableModel.

So, old transactions that are not being shown in the GUI deliberately are re appearing due the signaling update process.

This commits solves it, not showing transactions that appeared before the first tx time loaded into the model.
Ofc this is only applied if the wallet is exceeding the maximum loaded transactions limit in the GUI (20,000 transactions)

Github-Pull: #1749
Rebased-From: b711a7a26ace52f4d88a4eb980692c7635dc7964