wallet: Ignore MarkConflict if block hash is not known
>>> backports bitcoin/bitcoin@40e7b61835cbe5fd471d0b4b71972526bf0e523c

If number of conflict confirms cannot be determined, this means
that the block is still unknown or not yet part of the main chain,
for example during a reindex. Do nothing in that case,
instead of crash with an assertion.

Fixes bitcoin/bitcoin#7234.
Github-Pull: #1767
Rebased-From: f90fa087f52893b9d98a92e0ebb0b48f23081edb